% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBBtestfunction2.R
\name{JBBtestfunction2}
\alias{JBBtestfunction2}
\title{JBB test function 2}
\usage{
JBBtestfunction2(number)
}
\arguments{
\item{number}{Any number}
}
\value{
A message which depends on the input number
}
\description{
A trial function for a test package
}
\examples{
JBBtestfunction2(5)
JBBtestfunction2(42)
JBBtestfunction2(4.2)
}
